Added Install Script
Use this Arch Linux installation script with caution, as it is my first uploaded script and has not been thoroughly tested yet. However, with the help of ChatGPT, it's designed to make the installation process easy and efficient. If you encounter any issues or have any improvements, please submit them as commits or issues, and let me know about your testing experience.